[DAD-341] refactor: filter, handler 폴더 추가
