+++ src/main/java/com/forever/dadamda/config/SecurityConfig.java
@@ -1,6 +1,8 @@
 package com.forever.dadamda.config;
 
 import com.forever.dadamda.entity.Role;
+import com.forever.dadamda.filter.JwtAuthFilter;
+import com.forever.dadamda.handler.OAuth2SuccessHandler;
 import com.forever.dadamda.service.CustomOAuth2UserService;
 import com.forever.dadamda.service.TokenService;
 import lombok.RequiredArgsConstructor;
