+++ src/main/java/com/forever/dadamda/filter/JwtAuthFilter.java
@@ -1,4 +1,4 @@
-package com.forever.dadamda.config;
+package com.forever.dadamda.filter;
 
 import com.forever.dadamda.service.TokenService;
 import lombok.RequiredArgsConstructor;
